refactor: seats and rescheduleUId
